conns: Rearrange for generation
