+++ internal/conns/conns_test.go
@@ -9,80 +9,6 @@ import (
 	"github.com/hashicorp/aws-sdk-go-base/v2/servicemocks"
 )
 
-func TestAWSClientPartitionHostname(t *testing.T) {
-	testCases := []struct {
-		Name      string
-		AWSClient *AWSClient
-		Prefix    string
-		Expected  string
-	}{
-		{
-			Name: "AWS Commercial",
-			AWSClient: &AWSClient{
-				DNSSuffix: "amazonaws.com",
-			},
-			Prefix:   "test",
-			Expected: "test.amazonaws.com",
-		},
-		{
-			Name: "AWS China",
-			AWSClient: &AWSClient{
-				DNSSuffix: "amazonaws.com.cn",
-			},
-			Prefix:   "test",
-			Expected: "test.amazonaws.com.cn",
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.Name, func(t *testing.T) {
-			got := testCase.AWSClient.PartitionHostname(testCase.Prefix)
-
-			if got != testCase.Expected {
-				t.Errorf("got %s, expected %s", got, testCase.Expected)
-			}
-		})
-	}
-}
-
-func TestAWSClientRegionalHostname(t *testing.T) {
-	testCases := []struct {
-		Name      string
-		AWSClient *AWSClient
-		Prefix    string
-		Expected  string
-	}{
-		{
-			Name: "AWS Commercial",
-			AWSClient: &AWSClient{
-				DNSSuffix: "amazonaws.com",
-				Region:    "us-west-2", //lintignore:AWSAT003
-			},
-			Prefix:   "test",
-			Expected: "test.us-west-2.amazonaws.com", //lintignore:AWSAT003
-		},
-		{
-			Name: "AWS China",
-			AWSClient: &AWSClient{
-				DNSSuffix: "amazonaws.com.cn",
-				Region:    "cn-northwest-1", //lintignore:AWSAT003
-			},
-			Prefix:   "test",
-			Expected: "test.cn-northwest-1.amazonaws.com.cn", //lintignore:AWSAT003
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.Name, func(t *testing.T) {
-			got := testCase.AWSClient.RegionalHostname(testCase.Prefix)
-
-			if got != testCase.Expected {
-				t.Errorf("got %s, expected %s", got, testCase.Expected)
-			}
-		})
-	}
-}
-
 func TestGetSupportedEC2Platforms(t *testing.T) {
 	ec2Endpoints := []*servicemocks.MockEndpoint{
 		{
